fix ADMM name file, bnoise on KL divergence, and init
