+++ Wrappers/Python/test/test_SumFunction.py
@@ -81,8 +81,9 @@ class TestFunction(unittest.TestCase):
         M, N, K = 3,4,5
         ig = ImageGeometry(M, N, K)
         
-        tmp = ig.allocate('random_int', seed=1)
-        b   = ig.allocate('random_int', seed=2)
+        tmp = ig.allocate('random', seed=1)
+        b   = ig.allocate('random', seed=2)
+        eta = ig.allocate(0.1)
         
         operator = Identity(ig)
 
@@ -96,8 +97,8 @@ class TestFunction(unittest.TestCase):
         f7 = ZeroFunction()
         f8 = 5 *  ConstantFunction(10)             
         f9 = LeastSquares(operator, b, c=scalar)
-        f10 = 0.5*KullbackLeibler(b=b)
-        f11 = KullbackLeibler(b=b)
+        f10 = 0.5*KullbackLeibler(b=b,eta = eta)
+        f11 = KullbackLeibler(b=b, eta =eta)
         f12 = 10
         
 #        f10 = 0.5 * MixedL21Norm()
@@ -332,11 +333,10 @@ def test_ConstantFunction(self):
      
         
                 
-#if __name__ == '__main__':
+if __name__ == '__main__':
 #    
-#    t = TestFunction()
-#    t.test_SumFunction()
+    t = TestFunction()
+    t.test_SumFunction()
 #    t.test_SumFunctionScalar()
 
                 
-
